backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose
httpx
requests
fastapi-mail
python-multipart
email-validator
bson
pydantic
